fix: remove ghost deprecated test folder, optimize a slow test case
