+++ tests/integration_tests/operators/test_run.py
@@ -12,7 +12,7 @@ from airflow.utils.state import TaskInstanceState
 from wherobots.db import Region
 
 from airflow_providers_wherobots.operators.run import WherobotsRunOperator
-from tests.unit_tests.operator.test_run import build_ti
+from tests.unit_tests.operators.test_run import build_ti
 
 DEFAULT_START = pendulum.datetime(2021, 9, 13, tz="UTC")
 DEFAULT_END = DEFAULT_START + datetime.timedelta(days=1)
